move unicode superscript/subscript conversion to util; clean up type hints
